New Gear Meshes Added as well as Materials Ready for Engineers
I've added the gear meshes with a basic stone texture. It is ready for
the engineers to add collision boxes. and see if we can sync up the
teeth. the proper scale for the gear meshes is at 20% of original size,
so in the mesh settings it should always say .2 and should not be scaled
when placed in levels.