Set dressed E_4, E_6, E_8, E_10
Added stone detail in back to guide player. Also restructured some of
the background. Placed torches in background to assist in guiding the
player. Also moved background bricks slightly forward and back to create
depth. On E_8 (I think), I added an extra block, so that the player gear
doesn't just fly to the golden gear.