All of the Momentum and Water levels work
Didn't seem like the change did much. Otherwise I also put the second
background layer into all the water and xtreme levels. All levels should
also be in order of difficulty and the broken levels are at the end of
each level pack. WE HAVE 12 BROKEN LEVELS: 1 in momentum, 4 in water, 6
in buttons, and 1 of the xtreme levels.